take out tzip12 contracts and related files/tests
