--- contracts/fa2_multi_nft_faucet_storage.tz
@@ -1,1 +1,0 @@
-(Pair (Pair {} 0) (Pair {} {}))
