+++ contracts/ligo/fa2/fa2_interface.mligo
@@ -52,6 +52,7 @@ type update_operator =
   | Add_operator of operator_param
   | Remove_operator of operator_param
 
+(*
 type token_metadata =
 [@layout:comb]
 {
@@ -61,8 +62,16 @@ type token_metadata =
   decimals : nat;
   extras : (string, string) map;
 }
+*)
 
-type token_metadata_param = 
+type token_metadata =
+[@layout:comb]
+  {
+    token_id: token_id;
+    token_metadata_map: ((string, bytes) map);
+  }
+
+type token_metadata_param =
 [@layout:comb]
 {
   token_ids : token_id list;
@@ -73,7 +82,7 @@ type fa2_entry_points =
   | Transfer of transfer list
   | Balance_of of balance_of_param
   | Update_operators of update_operator list
-  | Token_metadata_registry of address contract
+  (* | Token_metadata_registry of address contract *)
 
 type fa2_token_metadata =
   | Token_metadata of token_metadata_param
