--- contracts/ligo/fa2/fa2_tzip16_interface.mligo
@@ -1,162 +1,0 @@
-#if ! FA2_INTERFACE
-#define FA2_INTERFACE
-
-type token_id = nat
-
-type transfer_destination =
-[@layout:comb]
-{
-  to_ : address;
-  token_id : token_id;
-  amount : nat;
-}
-
-type transfer =
-[@layout:comb]
-{
-  from_ : address;
-  txs : transfer_destination list;
-}
-
-type balance_of_request =
-[@layout:comb]
-{
-  owner : address;
-  token_id : token_id;
-}
-
-type balance_of_response =
-[@layout:comb]
-{
-  request : balance_of_request;
-  balance : nat;
-}
-
-type balance_of_param =
-[@layout:comb]
-{
-  requests : balance_of_request list;
-  callback : (balance_of_response list) contract;
-}
-
-type operator_param =
-[@layout:comb]
-{
-  owner : address;
-  operator : address;
-  token_id: token_id;
-}
-
-type update_operator =
-[@layout:comb]
-  | Add_operator of operator_param
-  | Remove_operator of operator_param
-
-(*
-type token_metadata =
-[@layout:comb]
-{
-  token_id : token_id;
-  symbol : string;
-  name : string;
-  decimals : nat;
-  extras : (string, string) map;
-}
-*)
-
-type token_metadata =
-[@layout:comb]
-  {
-    token_id: token_id;
-    token_metadata_map: ((string, bytes) map);
-  }
-
-type token_metadata_param =
-[@layout:comb]
-{
-  token_ids : token_id list;
-  handler : (token_metadata list) -> unit;
-}
-
-type fa2_entry_points =
-  | Transfer of transfer list
-  | Balance_of of balance_of_param
-  | Update_operators of update_operator list
-  (* | Token_metadata_registry of address contract *)
-
-type fa2_token_metadata =
-  | Token_metadata of token_metadata_param
-
-(* permission policy definition *)
-
-type operator_transfer_policy =
-  [@layout:comb]
-  | No_transfer
-  | Owner_transfer
-  | Owner_or_operator_transfer
-
-type owner_hook_policy =
-  [@layout:comb]
-  | Owner_no_hook
-  | Optional_owner_hook
-  | Required_owner_hook
-
-type custom_permission_policy =
-[@layout:comb]
-{
-  tag : string;
-  config_api: address option;
-}
-
-type permissions_descriptor =
-[@layout:comb]
-{
-  operator : operator_transfer_policy;
-  receiver : owner_hook_policy;
-  sender : owner_hook_policy;
-  custom : custom_permission_policy option;
-}
-
-(* permissions descriptor entrypoint
-type fa2_entry_points_custom =
-  ...
-  | Permissions_descriptor of permissions_descriptor contract
-
-*)
-
-
-type transfer_destination_descriptor =
-[@layout:comb]
-{
-  to_ : address option;
-  token_id : token_id;
-  amount : nat;
-}
-
-type transfer_descriptor =
-[@layout:comb]
-{
-  from_ : address option;
-  txs : transfer_destination_descriptor list
-}
-
-type transfer_descriptor_param =
-[@layout:comb]
-{
-  batch : transfer_descriptor list;
-  operator : address;
-}
-
-(*
-Entrypoints for sender/receiver hooks
-
-type fa2_token_receiver =
-  ...
-  | Tokens_received of transfer_descriptor_param
-
-type fa2_token_sender =
-  ...
-  | Tokens_sent of transfer_descriptor_param
-*)
-
-#endif
