+++ contracts/ligo/src/fa2_multi_nft_manager.mligo
@@ -4,13 +4,12 @@
 
 #include "fa2_multi_nft_token.mligo"
 
-type mint_token_param = 
+type mint_token_param =
 [@layout:comb]
 {
-  metadata : token_metadata;
+  token_metadata: token_metadata;
   owner : address;
 }
-
 
 type mint_tokens_param = mint_token_param list
 
@@ -19,20 +18,20 @@ type minted1 = {
   reversed_txs : transfer_destination_descriptor list;
 }
 
-let update_meta_and_create_txs (param, storage 
+let update_meta_and_create_txs (param, storage
     : mint_tokens_param * nft_token_storage) : minted1 =
-  let seed1 : minted1 = { 
-    storage = storage; 
+  let seed1 : minted1 = {
+    storage = storage;
     reversed_txs = ([] : transfer_destination_descriptor list);
   } in
   List.fold
     (fun (acc, t : minted1 * mint_token_param) ->
-      let new_token_id = t.metadata.token_id in
+      let new_token_id = t.token_metadata.token_id in
       if new_token_id < acc.storage.next_token_id
       then (failwith "FA2_INVALID_TOKEN_ID" : minted1)
       else
         let new_token_metadata =
-          Big_map.add new_token_id t.metadata acc.storage.token_metadata in
+          Big_map.add new_token_id t.token_metadata acc.storage.token_metadata in
 
         let new_storage = { acc.storage with
           token_metadata = new_token_metadata;
@@ -59,8 +58,9 @@ let mint_tokens (param, storage : mint_tokens_param * nft_token_storage)
     from_ = (None : address option);
     txs = mint1.reversed_txs;
   } in
-  let nop_operator_validator = 
+  let nop_operator_validator =
     fun (p : address * address * token_id * operator_storage) -> unit in
-  fa2_transfer ([tx_descriptor], nop_operator_validator, mint1.storage)
+  let ops, storage = fa2_transfer ([tx_descriptor], nop_operator_validator, mint1.storage) in
+  ops, storage
 
 #endif
